Do not catch `Throwable`, catch `Exception` instead.
Rationale:
Catching `Throwable` is discouradged as those object are not safe to catch. Objects that are safe to catch are `Exception` (and `AssertError`). The others (derivatives from `Error`) indicates that a D guarantee had failed and thus, the program is not in a valid state. Catching `Error` derivatives is the D runtime job, not ours.