Adapt build.d to require a dub instance
By making DUB self-hosted, we will allow a bunch of improvements,
most importantly using libraries and getting rid of the duplication
with the vibe.d code copied over.

However, in doing so, we want to make the process as easy for package manager as possible,
hence build.d will be left in place to provide user-friendly error messages.
The ones that wish to switch to a direct 'dub build' can do so, and the ones
that need backward compatibility can specify both HOST_DUB and DMD.